# [CANN] RoPE and CANCAT operator optimization